+++ pkg/manager/servicesLeader.go
@@ -14,8 +14,13 @@ import (
 	"k8s.io/client-go/tools/leaderelection/resourcelock"
 )
 
+// activeService keeps track of services that already have a leaderElection in place
+var activeService map[string]bool
+
 // The startServicesWatchForLeaderElection function will start a services watcher, the
 func (sm *Manager) startServicesWatchForLeaderElection(ctx context.Context) error {
+
+	activeService = make(map[string]bool)
 
 	err := sm.servicesWatcher(ctx, sm.startServicesLeaderElection)
 	if err != nil {
@@ -29,18 +34,23 @@ func (sm *Manager) startServicesWatchForLeaderElection(ctx context.Context) erro
 
 // The startServicesWatchForLeaderElection function will start a services watcher, the
 func (sm *Manager) startServicesLeaderElection(ctx context.Context, service *v1.Service, wg *sync.WaitGroup) error {
+	// No leader election is necessary
+	if activeService[string(service.UID)] {
+		return nil
+	}
+
 	id, err := os.Hostname()
 	if err != nil {
 		return err
 	}
 
 	serviceLease := fmt.Sprintf("kubevip-%s", service.Name)
-	log.Infof("Beginning services leadership, namespace [%s], lock name [%s], id [%s]", service.Namespace, serviceLease, id)
+	log.Infof("beginning services leadership, namespace [%s], lock name [%s], id [%s]", service.Namespace, serviceLease, id)
 	// we use the Lease lock type since edits to Leases are less common
 	// and fewer objects in the cluster watch "all Leases".
 	lock := &resourcelock.LeaseLock{
 		LeaseMeta: metav1.ObjectMeta{
-			Name:      plunderLock,
+			Name:      serviceLease,
 			Namespace: service.Namespace,
 		},
 		Client: sm.clientSet.CoordinationV1(),
@@ -48,7 +58,7 @@ func (sm *Manager) startServicesLeaderElection(ctx context.Context, service *v1.
 			Identity: id,
 		},
 	}
-
+	activeService[string(service.UID)] = true
 	// start the leader election code loop
 	leaderelection.RunOrDie(ctx, leaderelection.LeaderElectionConfig{
 		Lock: lock,
@@ -59,24 +69,27 @@ func (sm *Manager) startServicesLeaderElection(ctx context.Context, service *v1.
 		// get elected before your background loop finished, violating
 		// the stated goal of the lease.
 		ReleaseOnCancel: true,
-		LeaseDuration:   10 * time.Second,
-		RenewDeadline:   5 * time.Second,
-		RetryPeriod:     1 * time.Second,
+		LeaseDuration:   60 * time.Second,
+		RenewDeadline:   15 * time.Second,
+		RetryPeriod:     5 * time.Second,
 		Callbacks: leaderelection.LeaderCallbacks{
 			OnStartedLeading: func(ctx context.Context) {
-				err = sm.servicesWatcher(ctx, sm.syncServices)
-				if err != nil {
-					log.Error(err)
-				}
+				// we run this in background as it's blocking
+				go func() {
+					log.Infof("adding the service [%s/%s] with external address [%s]", service.Namespace, service.Name, service.Spec.LoadBalancerIP)
+					if err := sm.addService(service); err != nil {
+						log.Errorln(err)
+					}
+				}()
 			},
 			OnStoppedLeading: func() {
 				// we can do cleanup here
 				log.Infof("leader lost: %s", id)
-				for x := range sm.serviceInstances {
-					sm.serviceInstances[x].cluster.Stop()
+				if err := sm.deleteService(string(service.UID)); err != nil {
+					log.Errorln(err)
 				}
-
-				log.Fatal("lost leadership, restarting kube-vip")
+				// Mark this service is inactive
+				activeService[string(service.UID)] = false
 			},
 			OnNewLeader: func(identity string) {
 				// we're notified when new leader elected
